Prints elements in the list and return no of nodes

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -10,7 +10,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t num_elem = 0;
+	int num_elem = 0;
 
 	if (h == NULL)
 		return (num_elem);
